Don't rely on gitreviewers.init to build mailmap
Apparently, init build the mailmap, but it was empty by the time FindReviewers
ran. I'm not sure if that means init() has a different run order than I was
expecting or what, but this change is simpler. Mailmap wasn't used anywhere else
and its failure scenario isn't enough to stop the world.