Remove remaining references to git2go
NOTE: This removes a lot of tests too. Granted, the tests were useful to verify
certain logic, but they are incredibly stateful and have even resulted in loss
of work and files.

Once I come up with a way to mock out the filesystem and the git internals a bit
better, there will be a safer way to set up git repo scenarios and test against
them.