Handle cases where no eligible reviewers found
Any piece of code that hasn't been touched in 6+ months will have no reviewers
by default. This is because we use "recency" as a factor to influence experience
(the thought there being that even though I wrote some code, it might as well be
written by a stranger if enough time has passed).

This error prompts the user to use the "--since" argument to expand the scope of
eligible contributions to determine reviewers.